fixes BUG-47320: saved views display uuid for some FK filters
When a saved view included a filter by ID in a FK field that belongs to
a subtab that is not the first subtab of the root tab, the uuid instead
of the identifier was displayed.

This occurred because even the view was properly saved including the
current identifier for the filtered IDs, when it was applied it assumed
the filter editor was always already created to apply this cached
informantion. As tabs are lazyly created this is only true for the root
tab and its first child, but not for the rest. In these cases cache was
not applied displaying the uuid intead of the identifier.