Related to ISSUE-37322: Remove obsolete column c_projectline.dopricing
Earlier commit d918e6beee removed the ad_process_id=230 which was still
referenced by the ad_column_id=9873 which is a button (aka c_projectline.dopricing)

Remove that columnd and its (not displayed) field as they are useless
without a working process.
Reference a single reference to fhe field in C_PROJECT_WON pl-function
also which was just setting a default value for the column.