fixes ISSUE-47511: Entity.getProperty generates too much memory garbage
Entity.getProperty, which is invoked all the time, generated a huge
amount of memory garbage just preparing the message to be displayed in
case the requested property does not exist, which is thrown only very
rarely.

It has been optimized by moving the null check from Check.isNotNull to
Entity class and generating the message only when it is actually thrown.