Fixed ISSUE-47552: Invoice generator from goods shipment with positive tolerance
The fix makes sure we invoice any quantity shipped that wasn't been invoiced yet.

The code has been fixed in two concrete places, both associated with the scenario of
a goods shipment linked to a sales order:

1. When the invoice terms is After Order delivery: in this case que make sure we invoice
even in the case the order's delivery status is greater than 100 (that might happen with
quantity variable products or tolerances in WMS).

2. When the invoice terms allows to invoice before the shipment: in this case the formula
to calculate the qty to invoice has changed. Before it was:
  min(qtyOrdered - qtyInvoiced, qtyInThisShipment).
which was trying to support the corner case where the sales order was previously invoiced
before the shipment.

After the fix the formula is:
  qtyTotallyShipped - qtyPreviouslyInvoiced
and supports the corner case where the sales order has been previously invoiced but in a
different way. Instead of looking at the qty being shipped right now, it considers all the
shipped qty (including this shipment). This allows to easily know the qty actually pending
to invoice.