fixed BUG-47590: grid timeout not applied in audit trail deleted rows
Although correct grid query profile was set by fix 45a22d23973d for
BUG-44265, it was not applied when displaying deleted records. The
reason is deleted records queries are generated as manual sql and
executed through the ExecuteQuery API. This API was not taking into
account profile timeouts.

This commit:
 * Makes ExecuteQuery aware of query timeouts.
 * Allows AuditTrailDeletedRecords.getDeletedRecords to throw the
   exceptions that might occur during query execution so the caller
   catches them displaying a message to the user.
 * Removes an unused AuditTrailDeletedRecords.getDeletedRecords
   overloaded method.