Fixed ISSUE-47456: Order Line Attribute Set Value display logic
When the Product has associated a non-instanciable Attribute Set
without defining an Attribute Set Value and the Use Attribute Set Value As
fields, the purchase/sales order line's Attribute Set Value field had a
weird display logic behavior.
When creating a new order line, this field was not shown. However, it was
shown after saving the line and refreshing the record.

Problem was in the callout associated to the product, that was cleaning
some fields (inpattributeset and inpattrsetvaluetype) which are used in the
Attribute Set Value display logic.

The fix forces to clean these fields only when the attribute set is instanciable.
Note that in this case the field is set as read only.

In any other case these fields (inpattributeset and inpattrsetvaluetype) take the
values calculated in their respective Auxiliary inputs (ATTRIBUTESET and
ATTRSETVALUETYPE), and the Attribute Set Value display logic works as expected.

After the fix, the field's display logic is the same when creating a new record,
and after saving & refreshing it.