Fixes ISSUE-47793: Remove inactive report using jasper chart functionality in core
Jasperreports chart functionality will be moved to another module, and
ReportSalesPartnerProduct report was the only one using it in core. This
report has been long inactive, it was deactivated 9 years ago, as such,
it should be safe to remove and proceed with the separation of the chart
functinality libraries to another module.