Fixed ISSUE-47707: Modify Payment Plan wrong amounts autocalculated
Important: this is a partial fix that improves the situation in most
of the scenarios, but not in all. The platform used to validate and
generate new records is very weak and makes it impossible to properly
control all the scenarios. See #47798 and #47799 for details.

The main problem was while trying to retrieve the invoice header information.
The code was wrongly using "grid.view.parentWindow.activeView.getContextInfo"
which retrieves the Payment Plan record. Now we use
"grid.view.parentWindow.views[0].getContextInfo" to get the invoice header info.

totalExpected and totalReceived calculations were not used, so they are removed.

When adding a new record (OB.APRM.addNew) the calculation was doing:
outstandingAmt = Sum(GridExpected) - Sum(GridReceived) - Sum(GridOutstanding)

This algorithm won't work in the moment we delete a record in the grid.

Instead the algorithm now calculates the outstanding amount as:
outstandingAmt = invoiceOutstandingAmt - Sum(GridOutstandingAmt)

This is safer because it works with these scenarios:
1) If no payment plan is fully paid yet. Then all the records are displayed in the grid.
2) If there is any payment plan fully paid. Then those paid payment plans are not displayed.
3) If the user has changed any outstanding amount before creating the new record.

Then, the new record's expected amount is set equal to the calculated outstanding amount,
because it's a new record so both amount should always match.
If there are more adjustements in other lines' outstanding amount, the expected amounts will be
automatically recalculated.

The payment method information set in the new record is removed because the
"selectedRecord" object doesn't have this information. Instead, the payment method (and
any other information should be given in the moment of creating the new object by cloning
and existing one).