Fixes ISSUE-47773: Rename Sales Dimensional Report dimension from 'Ship to Address' to 'Partner Address'
Renamed the 'Ship to Address' dimension of the Sales Dimensional Report to 'Partner Address' as it reflects
better the report's behavior. The report groups by the order address, equivalent to the Business Partner
Address, so the 'Partner Address' name is more suitable than the 'Ship to Address' name for this dimension.