fixed BUG-48070: OOM compressing JS with certain regexp
When trying to compress a JS containing a matching for a single
character which includes the / character followed by the ` character an
OOM was produced. Ie. /[/`]/

This was caused by incorrectly detecting the / chracter inside the
single character matching group as regexp termination instead of as /
single character.

Note that other cases including other matching other quotes agfter /
character would have thrown an UnterminatedStringLiteralException.

This is fixed by taking into account when a regexp is inside a single
character match not to consider / as regexp termination.