Related to ISSUE-48259: M_Product_Category_Service.M_Product_ID child property
This column has been declared as Child Property in Parent Entity to have available a
method serviceProductCategoryList() in Product class, which is necessary to export this
info as part of the Product using the Retail API.

Note that the change makes sense from a functional point of view, especially considering
that the similar scenario for products [serviceProductList()] is already declared this way.

Besides, there shouldn't be a performance problem as the number of serviceProductCategory for
a product should be small in general.