Fixes ISSUE-48153: Validating error cleared when no mandatory fields are empty
To reproduce the issue, the user had to:
- create a record, filling in all mandatory fields
- empty one of the mandatory fields and save, an expected validation error was raised
- edit the line in grid view, mandatory field is given a default value (the FIC does it), but
after trying to save the validation error does not go away and the new value is not saved

The problem was that even though the FIC was providing a valid default value for the empty
mandatory field, the save process was not trying to save the record because it mistakingly
thought that the edit form had not been changed. To fix this, when processing the FIC return the
setHasChanged function is invoked if the FIC has provided a not-null value for a field that used
to be empty.