Fixes ISSUE-47617: Some language filters do not work properly
The problem was that in the datasource request to obtain the unique entries to populate the filter dropdown, the key property of that selector
was not being included in the _selectedProperties parameter.

In that request, _selectedProperties contained the properties of the grid the selector
belong to, but not the relevant properties of the referenced entity. For instance, in the request to populate the combo of the language selector
of User, the grid properties of User were included in _selectedProperties, but not "language" which is the key property of the referenced entity.

This is why in some windows worked, for instance the language combo in the Module window was properly populated because that window also contained
a "language" property, so by luck it had the same name as the key property of Language.

To fix this, we now ensure that they key property of the referenced entity is included in the _selectedProperties parameter. I prefer not to remove
the unrelated properties from _selectedProperties to be conservative and lower the regression risk.