Fixes ISSUE-48375: Grid fails refresh when an element is selected and a group by applied
Group by functionality doesn't generate a ResultSet when clicking the
refresh button, that means there is no data.resultSize object, it is
undefined. Because of this an error appeared when refreshing with an
element selected and with a group by applied to any column of a grid.

To fix it, the selected record check is skipped when doing a refresh
with a group by applied, as it is currently not supported to do so.
This basically results that when the refresh button is pressed, if a
record was selected(with group by applied), it will not be reselected
when the refresh finishes, if it is in the resultSet.