fixes BUG-48418: import entries with undeterministic keys can be processed twice
If an ImportEntryProcessor can return different keys for the same import
entry, ie. it returns a random number between 1 and 10 to distribute entries
in 10 threads, and an entry is takes long to be processed since it was
initially scheduled, it can be processed more than once.

This occurred because the mechanism to prevent processing an entry more
than once was checking the IDs of entries being processed only in the
assigend (by key) processor. If the key for the same entry varies among
different cycles, a different processor is assigned so the mechanism
did not detects the entry was already scheduled for processing.

Now the queu of processing entries has been moved from processor to
import entry manager, so there is a single centralized queue to perform
all the checks.