fixed BUG-48427: low prio entries can get precedence over high prio
Having scheduled in a given ImportEntryProcessRunnable a set of import
entries (all of them sharing same type of data and the same key). If the
processing of those entries does not finalize when the next ImportEntry
cycle looks for new entries to process and there are entries for the
same type of data and key, they will be scheduled to be executed in the
already running ImportEntryProcessRunnable. This makes these newly
scheduled entries to take precedence over new entries of other types of
data with a supposedly higher priority that require a new ImportEntryProcessRunnable.

Now ImportEntryProcessRunnable instances are linked to the
ImportEntryManager's execution cycle, so new entries for the same type
and key are not scheduled if there are still entries with the same key
and type scheduled from a previous cycle.