fixes BUG-48465: manage the DAL connection of the subreports
  Jasper subreports are filled with a multi-threading mechanism. In case
the threads use DAL to retrieve any information (for example an image)
we must ensure that the DAL connection is closed to prevent connection
leaks.

  For this reason, most of the Utility class methods are closing the
connection on finish. But at the same time this can cause problems in
case the DAL connection is reused both in the header and in the
subreport query. In that case if the header closes the DAL connection
through the Utility method, then the subreport will not be able to
execute the query because the connection is already closed.

  To fix this problems, two actions are being taken:

  - We provide our own subreport runner factory to ensure that the
threads used to fill the subreports close the DAL connections when they
are finished.

  - The Utility methods are no longer closing the DAL connection, because
this not needed anymore.