fixes ISSUE-48653: skip setImportEntryProcessed checks if invoked outside the IEM threads
  Improve the check that allows to set an import entry as processed. Now
apart from checking if the current node is in charge of handling the
import entries, it also checks if the thread setting the import entry
status is one of the ImportEntryManager threads.

  This is done now to avoid failing if this code is invoked outside of
the ImportEntryManager infrastructure, which is NOT recommended. In that
case now the "setImportEntryProcessed" method will not fail if is
executed in a node that is not the cluster leader.