Fixes ISSUE-48867: Process messages are not automatically removed
Messages in the message bar are automatically removed when some actions happen (i.e. when the contents of the current view are refreshed). When
a process is executed and its response contains a message we dont want that message automatically hidden, otherwise it will just blink away.

To fix it, we can set the keepOnAutomaticRefresh to true on the messageBar before displaying the message