[time-inv-cache] Change build loader method to a function instead of a CacheLoader
This is intended to reduce the exposed API from caffeine to the minimum.
All the API is now wrapped by the TimeInvalidatedCache.java and
TimeInvalidatedCacheBuilder.java and nothing should be exposed to the
outside, this way if there's any api change on their side, it would not
affect our implementation.