Fixes ISSUE-48939: Masterdata model that use extensions should take them into account when calculating the hash Masterdata model definition in the backend includes the list of properties that should be taken into account to calculate a hash. That hash will then be used to determine when those properties have changed, in order to force a full refresh.
Currently there are two methods that return the list of fields to take into account for the hash: getPropertiesFrom
- one that accepts the list of ModelExtension and that returns the list of all properties included in those extensions
- another one that accepts an entity and that returns all the properties of the given entity

The second one SHOULD NEVER be used in a master data model that uses ModelExtension, because if a new extension is included, or a new property in an existing extension is included, the new properties will not be included in the hash, and the full master data refresh that should take place will not happen.
To fix this, we search where the second method used and check if this module use the ModelExtension and after do the change.