Fixes ISSUE-49119: Prevents unneded pdf generation when printing multiple documents
When several records are selected and then clicking on the Print toolbar button, a single document is download,
and each document is attached to the records it belongs to if that option is chosen. During that process a copy of
each report is being stored in a temporary file, but apparently then that copy is not used.

Skipping the creation of those copies will save quite some time (50% estimated)