fixes BUG-49285: test connectivity process leaves an opened DB transaction
  In the test connectivity process we are making use of the thenApply
method of the CompletableFuture returned by the External System API to
build the process result with the information coming from the send
method.

  Altough we are consuming this API in a synchronous way, our lambda
executed in the thenApply is executed in a separate thread which is
making use of DAL to get the AD messages that are included in the
process result.

  For this reason we have to manually close the DAL connection
(transaction) in our lambda as this thread will not by controlled by the
standard flow which makes use of the DalThreadFilter to close the
connections.