Fixes ISSUE-49265: Remove obsolete Translation Required field from Module window
The problem that creating a new module with the translation required flag unchecked and create also a new window for that module after running the validatio it complains about that module not being marked as translation required.

To fix this, we removed the field translation required from Module window and also the function checkUIArtifacts from ModuleValidator. we deprecated the column in AD.