Fixes ISSUE-49398: Time property fields are properly displayed in form view
The TimeUIDefinition.getFieldProperties expects to receive the UTC value as a request parameter. When the field is
not a property field, the request parameter is named like this: "inp" + Sqlc.TransformaNombreColumna(field.getColumn().getDBColumnName()),
but when it is a property field the naming of the parameter is different. As a result, the current value was not found, so it was not
being converted from UTC to local time.

To fix it, the getFieldProperties method will first look for the standard parameter and if not found it will look for it using the property
field name convention