fixes BUG-49963: IEs not assigned to runnables are counted for idle time
After scheduling a batch of entries to be processed, the main IE threds
remains idle a time based on the count of scheduled entries.

In some cases, the entries obtained to be scheduled cannot finally be
set to a processing thread (ie. it was already full). These entries were
still counted unnecesarily increasing the idle waiting time.

Now we only count the entries that were actually scheduled.