Fixes ISSUE-50042: image columns can be used in ad_org table
The problem was that an event handler was taking for granted that all the tables it worked with were going to
include a property named "organization". This is true for all entities except for the Organization entity itself.

Now the event handler checks if the property exists before trying to retrieve it