Fixes ISSUE-50242: List property is properly updated in the status bar of the standard form
When a list formitem was updated by an event handler, its display value in a status bar was not being updated properly. The problem was that
the value used to determine what should be displayed was initially properly obtained using formItem.getDisplayValue, but then was incorrectly
overwritten by its identifier, that for this type of items is not relevant because the valueMap value should take precedence.

The change has been done in the getStatusBarFields that determines how the field is rendered in a status bar, the scope of the change is very
limited so the change should not be risky