fixes ISSUE-50721: can update database having merged modules
When updating database, it was not checked merged modules are not present as
sources. If a module is merged in another one, a validation should prevent
updating database until the merged module is uninstalled.

There was a buildvalidation to perform this, but it was partial. It has been
changed to:

  * Be generic, before only a set of "known" merges was checked.
  * Execute always, before only when updaing from <=17Q1 was executed