Fixes ISSUE-50874: ws request should not update inactive interval in standard sessions
Sessions created as a result of webservice requests are supposed to expire before standard sessions (1 minute vs 60 minutes by default). To force the shorter
expiration date of webservice sessions, the max inactive timeout was set to the smaller value each time a request to a webservice is received.

The problem is that if the request to the webservice was done as part of a standard session, we were making the standard session short lived as well, and that
was not supposed to happen.

To fix this, now the max inactive timeout is given a small value only if the session was created as a result of the webservice request