Fixes ISSUE-50921: Data cycles are now properly detected in TreeDatasourceService
The problem was that we were not properly checking the presence of data cycles. When filtering a node, we want to include also all its
ancestors. When looking for cycles, instead of looking for repeated nodes in the ancestor list, we were looking for repeated nodes in the
list of nodes that were going to be returned, that contains all the nodes that match the filter plus its ancestors.