Fixes ISSUE-51068: Block update from version older than 16Q1
Block updates coming from versions older than 16Q1 by adding a
buildvalidation to detect those cases and error out with a message.

Buildvalidation has an ExecutionLimit of 3.0.28207 which is exactly
16Q1.
ExecutionLimits are checked with < comparison which gives the wanted
behavior:
- Validation is running if old version is <16Q1
- But no longer running if it is already 16Q1 (or higher)

Additionally modify BuildValidationHandler declaring that this new check
should be run first before other validation:
- Conceptually makes sense to raise this first (before any data
  checks)
- Technically needed as in some (ratr) cases other buildvalidations
  already modify data (related to the upgrade process)
  i.e. RemoveCostingRuleProcessExecutions
  So check and block the update process as first step.
- ant target buildvalidation depends on core.lib which ensures the
  modified handler to be compiled before this new check is ran.

The buildvalidation committed .class file is compiled for Java11 with is
min version supported right now. For the upgrade process that is
already required anyway as i.e.:
- build.xml blocks version < Java 11 already
- dbsourcemanager.jar is also compiled for Java11
So that does not introduce any new restriction for those upgrades.

Wiki page is added with info for affected users and a table of added
restrictions (in case more come in future):
http://wiki.openbravo.com/wiki/RestrictOldDirectUpdates