Fixes ISSUE-51106: cannot download Excel report in Chrome
Excel is exported by opening a new tab/popup with a set target to
_blank.

This worked before chrome version 107, but it does not work anymore. The
main reason is that target=_blank implies that the new window will not
be able to access window.opener object(it will be null). To avoid this
problem, we set target=null when target is _blank, this allows us to
keep the old behaviour.