fix/51140: Fixes tree datasource node filtering
The problem we were having was that the tree datasource was looking for characteristics values with the name we provided in the combo, but not for our specific characteristic. Instead, it was looking for values for every characteristic, and that lead to a higher number of results than expected, reason that caused the problem

Because we need to set the whereClause of the datasource in order to build the new query, we have to make sure we obtained a fresh non-cached copy, to make sure the updated where clause does not spill in other flows.