Fixes ISSUE-52268: Update fileUpload library to its latest vesrion (1.5)
Update fileUpload library to the latest version (1.5) to use FileUploadBase#setFileCountMax.

The BaseProcessActionHandler class now uses the new setFileCountMax API to limit the number of file uploads it will support. The exact number is
obtained by checking how many upload parameters are defined for the given process. We need to include as part of the max number the number of
other parameters (such as processId, windowId, etc)