Fixes ISSUE-52859: The name of computed columns entities are not guaranteed to be unique.
The problem is that when generating the entities for the computed columns, we names of the entities are built using <className>_ComputedColumns, and the name for both m_product and m_offer_product classes is Product, resulting in two entities with the same name (Product_ComputedColumn).

To fix it, the name of the entities generated for the computed columns, the original entity name (guaranteed to be unique) will be used as prefix instead of the class name (not guaranteed).

Note that this change is an API change due to the name of the computed column entities being updated.