Fixes ISSUE-52851: Invalid Role error accesing to WebPOS with an User with no Backend access
An User with a single active Role with restricted backend access should not be able to Log in into the backend but it should be able to access other applications such as the WebPOS. However, when attempting to log in into WebPOS the request returns the following error: "No valid Role identified. Please contact your system administrator for access".

The problem here that withe the restricted backend acces we are not checking if the role is active or not.

So we added this part in the function checkLicenseAndGo in the LoginHandler.