Fixes ISSUE-53228: Variant product with custom images can be deleted
The problem was that the check being done in RemoveImagesEventHandler to determine if a given image was still
referenced by other records was not done properly. The check was always being done against the product.ad_image_id
column, even though the image property being checked could be another.

As part of this fix, also the hardcoded reference to Product has been updated, to make sure this event handler works
with other entities that contain product images