Related to ISSUE-53450: Add a language column to ad_org
Add a new field to the organization window to be able to configure the language login page in POS and POS2.

- Create a new column "language" in the table "ad_org" with a foreign key constraint refence to the table "ad_language".
- Create a new field "language" on "organization" window.

The language is added in this module and not in retail.posterminal because the language field in the Organization table seems useful
enough to be used by other modules.