Fixes ISSUE-53514: Block build when trying to use Oracle Database
With the 23Q4 release we stop supporting to run Openbravo on top of
Oracle Database.
To avoid that being missed this adds a check to detect and fail with a
clear message.

As that should trigger on any task (install and updates) the check is
added to the common "init" target running before all of all of them.

That target also has a range of similar checks already (i.e. JAVA_HOME,
min version for java and ant)