fixes BUG-53535: error updating records referencing external system
The ExternalSystemDataEventHandler was intended to invalidate cached
configurations that may have changed. For example, for HTTP based
external systems, these are changes done in the HTTP subtab.

But this event handler should not be fired when changing records of
other entities that has a reference to an external system. For this
reason we have added the isParent check because the configuration
entities are represented as children of the ExternalSystemData entity
and the rest of entities should not.

This changeset also renames the constant as EXTERNAL_SYSTEM_PROPERTY
to follow the standard convention.