Fixes ISSUE-53659: Remove 2 ancient dal junit tests
Those 2 test classes did not really test anything. But were rather code
example on how to convert some non-trivial SQL query piece by piece to
HQL.

That is more a documentation topic and both Howto's are still present
on the wiki).
Additionally those code-examples where not updated i.e. with HQL style
so are no best practice to use as is.

Remove the 2 unit tests classes.